refactor: socials util
