+++ the-name-card/src/components/WidgetItem/Items/index.ts
@@ -1,3 +1,4 @@
 export * from './itemsUtil';
+export * from './socialsUtil';
 export * from './Socials';
 export * from './Link';
